rename molecule to mol
